Move all packages from /pkg to /internal
Ahead of an initial 1.0.0 release, we shouldn't preemptively
expose any packages publicly. Instead, let's start my making
any packages internal, and we can expose them publicly
later on. This commit moves everything currently in /pkg
to /internal for now.
